Do better with punctuation.
Handle things like screambot scream! or screambot scream??
Also handle it when we get the text "@screambot" instead of the uid.
And stop using 'string' as a variable name, come on.
And upgrade to python3 at last.

These should all be separate commits and PRs obv but I have to go to work in a
minute so I'm shipping it!